Moved magnet out of its own package
